+++ test/Armada.Test.Unit/Program.cs
@@ -256,6 +256,7 @@ namespace Armada.Test.Unit
             runner.AddSuite(new WorkflowProfileCheckRunServiceTests());
             runner.AddSuite(new CheckRunIsolatedCheckoutTests());
             runner.AddSuite(new SiblingLeaseRegistryTests());
+            runner.AddSuite(new ConsumerVesselResolverTests());
             runner.AddSuite(new DiskLifecycleTests());
             runner.AddSuite(new CheckRunParsingServiceTests());
             runner.AddSuite(new MergeRefCandidateTests());